Move create exercise from tools to bin
